+++ frontend/src/api/actions/ddahs.js
@@ -14,7 +14,7 @@ import {
     arrayToHash,
     flattenIdFactory,
 } from "./utils";
-import { apiGET, apiPOST } from "../../libs/apiUtils";
+import { apiGET, apiPOST } from "../../libs/apis";
 import { createSelector } from "reselect";
 import { activeRoleSelector } from "./users";
 import { ddahsReducer } from "../reducers";
