+++ frontend/src/api/actions/offers.js
@@ -9,7 +9,7 @@ import {
 } from "../constants";
 import { fetchError } from "./errors";
 import { actionFactory, validatedApiDispatcher } from "./utils";
-import { apiGET, apiPOST } from "../../libs/apiUtils";
+import { apiGET, apiPOST } from "../../libs/apis";
 import { fetchAssignment } from "./assignments";
 import { activeRoleSelector } from "./users";
