+++ frontend/src/libs/import-export/index.js
@@ -1,8 +1,8 @@
-export * from "./prepareData";
-export * from "./spreadsheetRowMapper";
-export * from "./dataToFile";
-export * from "./matchByUtoridOrName";
-export * from "./normalizeImport";
+export * from "./prepare-data";
+export * from "./spreadsheet-row-mapper";
+export * from "./data-to-file";
+export * from "./match-by-utorid-or-name";
+export * from "./normalize-import";
 export * from "./validate";
-export * from "./prepareSpreadsheet";
-export * from "./prepareJson";
+export * from "./prepare-spreadsheet";
+export * from "./prepare-json";
