+++ frontend/src/views/positions/import-export.js
@@ -16,7 +16,7 @@ import {
     PositionsList,
     PositionsDiffList,
 } from "../../components/positions-list";
-import { diffImport, getChanged } from "../../libs/diffUtils";
+import { diffImport, getChanged } from "../../libs/diffs";
 
 /**
  * Allows for the download of a file blob containing the exported instructors.
